+++ src/compas/datastructures/mesh/mesh.py
@@ -473,8 +473,13 @@ class Mesh(HalfEdge):
 
         return mesh
 
-    def to_vertices_and_faces(self):
+    def to_vertices_and_faces(self, triangulated=False):
         """Return the vertices and faces of a mesh.
+
+        Parameters
+        ----------
+        triangulated: bool, optional
+            If True, triangulate the faces.
 
         Returns
         -------
@@ -486,7 +491,31 @@ class Mesh(HalfEdge):
         """
         key_index = self.key_index()
         vertices = [self.vertex_coordinates(key) for key in self.vertices()]
-        faces = [[key_index[key] for key in self.face_vertices(fkey)] for fkey in self.faces()]
+
+        if not triangulated:
+            faces = [[key_index[key] for key in self.face_vertices(fkey)] for fkey in self.faces()]
+            return vertices, faces
+
+        faces = []
+
+        for fkey in self.faces():
+            face_vertices = self.face_vertices(fkey)
+
+            if len(face_vertices) == 3:
+                a, b, c = face_vertices
+                faces.append([key_index[a], key_index[b], key_index[c]])
+            elif len(face_vertices) == 4:
+                a, b, c, d = face_vertices
+                faces.append([key_index[a], key_index[b], key_index[c]])
+                faces.append([key_index[a], key_index[c], key_index[d]])
+            else:
+                centroid = centroid_polygon([vertices[key_index[key]] for key in face_vertices])
+                ckey = len(vertices)
+                vertices.append(centroid)
+
+                for a, b in pairwise(face_vertices + face_vertices[:1]):
+                    faces.append([key_index[a], key_index[b], ckey])
+
         return vertices, faces
 
     @classmethod